ai-service/requirements.txt
flask
numpy
scikit-learn
joblib
opencv-python
kafka-python
tensorflow